src/requirements.txt
mediapipe>=0.10.9
numpy>=1.22.0
opencv-python>=4.9.0.80
Pillow